Floodgate (Bedrock) support & a name bypass list
This update adds a Floodgate API provider, allowing the server to enable/disable if they want Bedrock Players to be checked for a VPN. This is because bedrock can cause more cases of false detections due to factors such as cellular data being used.

I have also added a bypass-name list, allowing the server owner to bypass certain account names as a secondary option to the permission system.